# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: reports
    description: Pre-built, read-only report library (server-computed analytics)
  - name: lookup
    description: Ad-hoc, read-only Bullhorn lookups
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /v1/reports:
    get:
      operationId: listReports
      tags: [reports]
      summary: List available reports
      description: Returns the catalog of pre-built reports and their parameters.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Report catalog
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/reports/staffing-scorecard:
    get:
      operationId: getStaffingScorecard
      tags: [reports]
      summary: Staffing scorecard
      description: >-
        Year-to-date staffing scorecard by department: confirmed placements
        (Contract / Contract-to-Hire / Direct Hire), open jobs, active
        opportunities, and a demand-vs-delivery ratio.
      security:
        - bearerAuth: []
      parameters:
        - name: year
          in: query
          required: false
          description: Calendar year; defaults to the current year.
          schema:
            type: integer
            minimum: 2000
            maximum: 2100
      responses:
        "200":
          description: Scorecard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/reports/placements:
    get:
      operationId: getPlacementsReport
      tags: [reports]
      summary: Placements report
      description: Confirmed placements over a period, by department and employment type.
      security:
        - bearerAuth: []
      parameters:
        - name: startDate
          in: query
          required: false
          description: Inclusive start date (YYYY-MM-DD). Defaults to the start of the current year.
          schema:
            type: string
            pattern: '^\d{4}-\d{2}-\d{2}$'
            example: "2026-01-01"
        - name: endDate
          in: query
          required: false
          description: Inclusive end date (YYYY-MM-DD). Defaults to today.
          schema:
            type: string
            pattern: '^\d{4}-\d{2}-\d{2}$'
            example: "2026-06-30"
        - name: status
          in: query
          required: false
          description: "'confirmed' (default) or 'all'."
          schema:
            type: string
            enum: [confirmed, all]
      responses:
        "200":
          description: Placements report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/reports/open-jobs:
    get:
      operationId: getOpenJobsReport
      tags: [reports]
      summary: Open jobs / demand report
      description: Current open requisitions by department and by employment type.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Open jobs report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/reports/sales-pipeline:
    get:
      operationId: getSalesPipelineReport
      tags: [reports]
      summary: Sales pipeline report
      description: Active sales opportunities by department and by stage.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Sales pipeline report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/reports/job-aging:
    get:
      operationId: getJobAgingReport
      tags: [reports]
      summary: Job aging report
      description: Open requisitions bucketed by age, with stale (>90 days) reqs by department.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Job aging report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/reports/recruiter-leaderboard:
    get:
      operationId: getRecruiterLeaderboard
      tags: [reports]
      summary: Recruiter leaderboard
      description: >-
        Recruiters ranked by submission-to-placement conversion. Soft walls are
        channel realism — never a dead end. On wall_time follow
        asyncContinuation.rest: POST /v1/reports/recruiter-leaderboard/jobs then
        poll GET /v1/reports/jobs/{jobId}.
      security:
        - bearerAuth: []
      parameters:
        - name: startDate
          in: query
          required: false
          description: Inclusive start date (YYYY-MM-DD). Defaults to the start of the current year.
          schema:
            type: string
            pattern: '^\d{4}-\d{2}-\d{2}$'
        - name: endDate
          in: query
          required: false
          description: Inclusive end date (YYYY-MM-DD). Defaults to today.
          schema:
            type: string
            pattern: '^\d{4}-\d{2}-\d{2}$'
      responses:
        "200":
          description: Recruiter leaderboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/reports/recruiter-leaderboard/jobs:
    post:
      operationId: startRecruiterLeaderboardJob
      tags: [reports]
      summary: Start async recruiter leaderboard job
      description: >-
        Soft-wall continuation. Returns HTTP 202 with jobId. Poll
        GET /v1/reports/jobs/{jobId} until status=complete|failed.
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                startDate:
                  type: string
                endDate:
                  type: string
      responses:
        "202":
          description: Job accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/reports/scout-qualified-by-department:
    get:
      operationId: getScoutQualifiedByDepartment
      tags: [reports]
      summary: Scout Screen qualified by department
      description: >-
        Screening/Scout notes by Internal Department — natural-language ready.
        Resolves nicknames (STSI→STS-STSI). For 'list N most recent' pass
        limit=N. Returns stopReason + confirmedComplete + asyncContinuation —
        soft walls are channel realism, never a dead end. On wall_time follow
        asyncContinuation.rest: POST /v1/reports/scout-qualified-by-department/jobs
        then poll GET /v1/reports/jobs/{jobId} (same args). Keep working unless
        confirmedComplete or a real connector/gateway limit. Never date-window
        fan-out. Never give up solely because of wall_time.
      security:
        - bearerAuth: []
      parameters:
        - name: department
          in: query
          required: true
          schema:
            type: string
        - name: noteAction
          in: query
          required: false
          schema:
            type: string
        - name: openJobsOnly
          in: query
          required: false
          schema:
            type: boolean
        - name: applicantPool
          in: query
          required: false
          schema:
            type: string
            enum: [responses, all]
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
        - name: mode
          in: query
          required: false
          schema:
            type: string
            enum: [bounded, exhaustive]
        - name: maxJobs
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 2000
        - name: maxCandidatesToScan
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 800
        - name: dateAddedStart
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: dateAddedEnd
          in: query
          required: false
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Scout qualified candidates by department
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/reports/scout-qualified-by-department/jobs:
    post:
      operationId: startScoutDeptReportJob
      tags: [reports]
      summary: Start async scout department report job
      description: >-
        Universal soft-wall continuation for Actions/REST hosts. Accepts the
        same body fields as the sync scout report query params. Returns HTTP
        202 with jobId. Poll GET /v1/reports/jobs/{jobId} until
        status=complete|failed. Use when sync GET returns stopReason=wall_time /
        asyncContinuation (same args; ~20 min safety max). Never date-window
        fan-out. Never give up solely because of the soft wall.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [department]
              properties:
                department:
                  type: string
                noteAction:
                  type: string
                openJobsOnly:
                  type: boolean
                applicantPool:
                  type: string
                  enum: [responses, all]
                limit:
                  type: integer
                  minimum: 1
                  maximum: 50
                mode:
                  type: string
                  enum: [bounded, exhaustive]
      responses:
        "202":
          description: Job accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/reports/jobs/{jobId}:
    get:
      operationId: getReportJob
      tags: [reports]
      summary: Poll async report job
      description: >-
        Universal soft-wall poll for Actions/REST hosts. Firm-scoped job status
        (queued|running|complete|failed). When status=complete, includes the
        tool result (scout, match_candidates_for_job, or recruiter_leaderboard).
        Keep polling after sync wall_time — never give up on the soft wall alone.
      security:
        - bearerAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Report job status / result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Job not found for this firm
  /v1/sourcing/match-candidates-for-job:
    post:
      operationId: matchCandidatesForJob
      tags: [sourcing]
      summary: Match candidates for a job
      description: >-
        Ranked, résumé-verified shortlist for a JobOrder. Soft walls are channel
        realism — never a dead end. On wall_time follow asyncContinuation.rest:
        POST /v1/sourcing/match-candidates-for-job/jobs then poll
        GET /v1/reports/jobs/{jobId}.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [jobId]
              properties:
                jobId:
                  type: integer
                mustHaveSkills:
                  type: array
                  items:
                    type: string
                niceToHaveSkills:
                  type: array
                  items:
                    type: string
                limit:
                  type: integer
                  minimum: 1
                  maximum: 15
                poolSize:
                  type: integer
                  minimum: 1
                  maximum: 100
                localOnly:
                  type: boolean
                includePlaced:
                  type: boolean
                includeSubmitted:
                  type: boolean
                includeDoNotContact:
                  type: boolean
                includeInactive:
                  type: boolean
      responses:
        "200":
          description: Match result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/sourcing/match-candidates-for-job/jobs:
    post:
      operationId: startMatchCandidatesJob
      tags: [sourcing]
      summary: Start async match-candidates job
      description: >-
        Soft-wall continuation. Same body as sync match. Returns HTTP 202 with
        jobId. Poll GET /v1/reports/jobs/{jobId} until complete|failed.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [jobId]
              properties:
                jobId:
                  type: integer
                mustHaveSkills:
                  type: array
                  items:
                    type: string
                niceToHaveSkills:
                  type: array
                  items:
                    type: string
                limit:
                  type: integer
                  minimum: 1
                  maximum: 15
                poolSize:
                  type: integer
                  minimum: 1
                  maximum: 100
                localOnly:
                  type: boolean
                includePlaced:
                  type: boolean
                includeSubmitted:
                  type: boolean
                includeDoNotContact:
                  type: boolean
                includeInactive:
                  type: boolean
      responses:
        "202":
          description: Job accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/count:
    post:
      operationId: countEntities
      tags: [lookup]
      summary: Count Bullhorn records
      description: >-
        Exact count for a Lucene query, optionally broken down by a field.
        Read-only; returns totals without the underlying records.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CountRequest"
      responses:
        "200":
          description: Count result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    CountRequest:
      type: object
      properties:
        entityType:
          type: string
          description: >-
            Searchable entity: Candidate, ClientContact, ClientCorporation,
            JobOrder, JobSubmission, Placement, Lead, Opportunity, or Note.
        query:
          type: string
          description: Lucene query string. Omit to count all records of the entity.
        groupBy:
          type: string
          description: Field to break the count down by (e.g. status, correlatedCustomText1).
        groupValues:
          type: array
          items:
            type: string
          description: Known values for groupBy, for an exact (non-sampled) breakdown.
      required:
        - entityType
    ReportResult:
      type: object
      description: Report or count payload. The exact shape varies by endpoint.
      additionalProperties: true